qa: fix how we attempt to install apache2
There's a bug in the apache2 package whereby one hook writes to the /etc/apache2/sites-available/000-default.conf file, and then a later one attempts to do so as well, thus creating a conflict situation.
Providing the `-o Dpkg::Options::="--force-confnew"` option ensures the new version from the later hook is installed.
In the end, it doesn't matter which one is installed, because we write over it, but without this, the installation fails.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>